create internal/check package for ClusterClass
